+++ client/app/models/paper.js
@@ -10,6 +10,9 @@ const PAPER_SUBMITTABLE_STATES = [
   'invited_for_full_submission'
 ];
 
+const TERMINAL_STATES = ['accepted', 'rejected'];
+const DECIDABLE_STATES = ['submitted', 'initially_submitted', 'checking'];
+
 const PAPER_GRADUAL_ENGAGEMENT_STATES = [
   'unsubmitted',
   'initially_submitted', // different than submittable states
@@ -42,7 +45,8 @@ export default DS.Model.extend({
   body: attr('string'),
   coverEditors: attr(),
   createdAt: attr('date'),
-  creator: belongsTo('user', { async: true }),
+  creator: belongsTo('user', { async: false }),
+  shortDoi: attr('string'),
   doi: attr('string'),
   editable: attr('boolean'),
   editorMode: attr('string', { defaultValue: 'html' }),
@@ -68,6 +72,7 @@ export default DS.Model.extend({
   withdrawalReason: attr('string'),
   url: attr('string'),
 
+  paper_shortDoi: computed.oneWay('shortDoi'),
   allAuthorsUnsorted: computed.union('authors', 'groupAuthors'),
   allAuthorsSortingAsc: ['position:asc'],
   allAuthors: computed.sort('allAuthorsUnsorted', 'allAuthorsSortingAsc'),
@@ -94,8 +99,8 @@ export default DS.Model.extend({
     return this.get('roles').sort().join(', ');
   }),
 
-  latestDecision: computed('decisions.@each.latest', function() {
-    return this.get('decisions').findBy('latest', true);
+  draftDecision: computed('decisions.@each.draft', function() {
+    return this.get('decisions').findBy('draft', true);
   }),
 
   latestRegisteredDecision: computed(
@@ -107,7 +112,7 @@ export default DS.Model.extend({
 
   previousDecisions: computed('decisions.@each.registeredAt', function() {
     return this.get('decisions')
-      .filterBy('registeredAt')
+      .rejectBy('draft')
       .sortBy('registeredAt')
       .reverseObjects();
   }),
@@ -176,6 +181,11 @@ export default DS.Model.extend({
 
   isInitialSubmission: computed.and('gradualEngagement', 'isUnsubmitted'),
   isFullSubmission: computed.and('gradualEngagement', 'invitedForFullSubmission'),
+
+  /* True if a decision can be registered in this state. */
+  isReadyForDecision: computed('publishingState', function() {
+    return DECIDABLE_STATES.includes(this.get('publishingState'));
+  }),
 
   engagementState: computed('isInitialSubmission', 'isFullSubmission', function(){
     if (this.get('isInitialSubmission')) {