+++ client/tests/components/author-form-test.js
@@ -2,6 +2,7 @@ import {moduleForComponent, test} from 'ember-qunit';
 import FactoryGuy from 'ember-data-factory-guy';
 import { manualSetup } from 'ember-data-factory-guy';
 import { createQuestionWithAnswer } from 'tahi/tests/factories/nested-question';
+import TestHelper from 'ember-data-factory-guy/factory-guy-test-helper';
 
 import hbs from 'htmlbars-inline-precompile';
 
@@ -14,23 +15,30 @@ moduleForComponent(
       manualSetup(this.container);
 
       $.mockjax({url: '/api/countries', status: 200, responseText: {
-        countries: [],
+        countries: []
       }});
       $.mockjax({url: '/api/institutional_accounts', status: 200, responseText: {
-        institutional_accounts: [],
+        institutional_accounts: []
       }});
+
+      let journal = FactoryGuy.make('journal');
+      TestHelper.mockFind('journal').returns({model: journal});
 
       let user = FactoryGuy.make('user');
       let task = FactoryGuy.make('authors-task');
       let author = FactoryGuy.make('author', { user: user });
+      let paper = FactoryGuy.make('paper');
 
       this.set('author', author);
+      this.set('author.paper', paper);
+      this.set('author.paper.journal', 1);
       this.set('isNotEditable', false);
       this.set('model', Ember.ObjectProxy.create({object: author}));
       this.set('task', task);
 
       this.set("toggleEditForm", () => {});
       this.set("validateField", () => {});
+      this.set("canRemoveOrcid", true);
 
       createQuestionWithAnswer(author, 'author--published_as_corresponding_author', true);
       createQuestionWithAnswer(author, 'author--deceased', false);
@@ -61,6 +69,7 @@ var template = hbs`
       hideAuthorForm="toggleEditForm"
       isNotEditable=isNotEditable
       saveSuccess=(action toggleEditForm)
+      canRemoveOrcid=true
   }}`;
 
 test("component displays the orcid-connect component when the author has an orcidAccount", function(assert){