+++ spec/controllers/orcid_accounts_controller_spec.rb
@@ -30,46 +30,6 @@ describe OrcidAccountsController do
         end
 
         it { is_expected.to responds_with(200) }
-
-        describe "oauth_authorize_url in the JSON response" do
-          let(:oauth_authorize_url) { res_body['orcid_account']['oauth_authorize_url'] }
-          let(:redirect_uri) { 'http://test.host/api/orcid/oauth' }
-
-          it "hits the ORCID server" do
-            do_request
-            expect(oauth_authorize_url).to match(/#{TahiEnv.orcid_site_host}/)
-          end
-
-          it "hits /oauth/authorize" do
-            do_request
-            expect(oauth_authorize_url).to match(%r{/oauth/authorize})
-          end
-
-          it "contains the ORCID_KEY" do
-            do_request
-            expect(oauth_authorize_url).to match(/#{TahiEnv.orcid_key}/)
-          end
-
-          it "passes a response_type of 'code'" do
-            do_request
-            expect(oauth_authorize_url).to match(/response_type=code/)
-          end
-
-          it "requests a scope of '/read-limited'" do
-            do_request
-            expect(oauth_authorize_url).to match(%r{scope=/read-limited})
-          end
-
-          it "requests only one scope" do
-            do_request
-            expect(oauth_authorize_url).not_to match(%r{scope=/[\w-]*%20/})
-          end
-
-          it "passes in a redirect uri" do
-            do_request
-            expect(oauth_authorize_url).to match(/redirect_uri=#{redirect_uri}/)
-          end
-        end
       end
 
       context 'when ORCID_CONNECT_ENABLED is false' do
@@ -91,9 +51,10 @@ describe OrcidAccountsController do
 
     it_behaves_like 'an unauthenticated json request'
 
-    context 'when the user is signed in' do
+    context 'when the user is signed in with remove_orcid permission' do
       before do
         stub_sign_in(user)
+        allow(user).to receive(:can?).with(:remove_orcid, Journal).and_return(true)
       end
 
       context 'when ORCID_CONNECT_ENABLED is true' do
@@ -113,6 +74,36 @@ describe OrcidAccountsController do
           do_request
           serializer = orcid_account.active_model_serializer.new(orcid_account, scope: orcid_account)
           expect(res_body.keys).to match_array(serializer.as_json.stringify_keys.keys)
+        end
+      end
+
+      context 'when ORCID_CONNECT_ENABLED is false' do
+        around do |example|
+          ClimateControl.modify(ORCID_CONNECT_ENABLED: 'false') do
+            example.run
+          end
+        end
+
+        it { is_expected.to responds_with(404) }
+      end
+    end
+
+    context 'when the user does not have remove_orcid permission' do
+      before do
+        stub_sign_in(user)
+        allow(user).to receive(:can?).with(:remove_orcid, Journal).and_return(false)
+      end
+
+      context 'when ORCID_CONNECT_ENABLED is true' do
+        around do |example|
+          ClimateControl.modify(ORCID_CONNECT_ENABLED: 'true') do
+            example.run
+          end
+        end
+
+        it 'is not authorized' do
+          do_request
+          expect(response.status).to eq(403)
         end
       end